chemo-pharmacy-backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
alembic==1.15.2
pydantic==2.11.4
pydantic-settings==2.9.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
pytest==8.3.5
httpx==0.28.1
